fix: surface-ref-block interaction
